requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-openai
unstructured==0.16.10
fastapi
uvicorn
python-dotenv
langchain-google-genai
faiss-cpu
Markdown
streamlit

-e .